Pull prefixes out of rdflib Graph for use in serialisation
ns_prefix_dict extracts the namespace-prefix pairs from an rdflib
Graph and returns them in a dictionary.

This is useful for those who want to use the serialisation for rdflib
Graphs that they know are SBOL graphs, and which they have constructed
with specific prefixes that they want to include in the
serialisation.